Add new `floating` (BOOL) property.
This property allows the NoticeView to be floating at the same position
in a UIScrollView, while the user is scrolling. It's pretty much the
same implementation as in the pull request #41, using KVO. It's just
based upon the most recent HEAD of master, and includes some fix
(leaked observers). There're also two new buttons in the example project
to test it.