Added the option to opt out of a dismissal in the dismissal block.
Added ``-setDismissalBlockWithOptionalDismiss:`` that accepts a second block attribute that is a BOOL pointer. You can then opt out from dismissal of the ``NoticeView`` from the block. This is good when you want the user to be able to, for example, tap a sticky ``NoticeView``, get an ``AlertView`` that wants a confirmation from the user and depending on the outcome dismiss the ``NoticeView`` or not.

Usage
```objective-c
[self.noticeView setDismissalBlockWithOptionalDismiss:^(BOOL dismissedInteractively, BOOL *dismissAfterBlock) {
    *dismissAfterBlock = YES; // or NO if you don't want the NoticeView to be dismissed
}];
```